pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "test-warden"
version = "0.1.0"
description = "AI-powered auto-healing for Selenium/Playwright test suites"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Test Warden Team" }
]
keywords = ["testing", "selenium", "playwright", "ai", "auto-healing"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
]

dependencies = [
    # CLI
    "click>=8.1.0",
    "rich>=13.0.0",
    
    # HTML/DOM Analysis
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    
    # AI/LLM
    "google-generativeai>=0.8.0",
    "langgraph>=0.2.0",
    "langfuse>=2.0.0",
    
    # Config
    "pyyaml>=6.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    
    # Code parsing
    "tree-sitter>=0.21.0",
    "tree-sitter-python>=0.21.0",
    "tree-sitter-javascript>=0.21.0",
    
    # Utilities
    "httpx>=0.27.0",
    "tenacity>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
]

[project.scripts]
test-warden = "test_warden.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/test_warden"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
